move romfs and lz tool to common
